+++ components/Player/DefilingText.tsx
