+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import Header from '../components/Header'
-import Sidebar from '../components/Sidebar'
+import Sidebar from '../components/Sidebar/Index'
 import '../styles/globals.css'
 
 function MyApp({ Component, pageProps }) {
